record: Wait for the spawned target process
When timeout is not given and the target is command, 'record' does not
wait for the opened pipe, so the target process became defunct.  As a
result, kdamond does not finished and the 'record' also do not finished.
This commit fixes the problem by waiting on the pipe with the timeout
argument.

Fixes: c951a00e316f ("record: Implement 'timeout' option")
Signed-off-by: SeongJae Park <sjpark@amazon.de>