record: Make the --output_permission applied from the beginning
Because --output_permission is applied at the end of record, users could
see the output file of different permission in the middle of execution.
It's inevitable for kernels having record feature.  This commit avoids
it for kernels having no record feature by specifying the permission
when opening the output file.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>