_damon: Remove 'damon_{debug,sys}fs' features in supported features list
'damon_debugfs' and 'damon_sysfs' features in the supported features
list is not properly updated as features update is done depending on
selected DAMON interface.  Further, users can know if DAMON debugfs
and/or sysfs interfaces are supported with 'damo-features' with
'--damon_interface' option.  This commit therefore remove the
'damon_debugfs' and 'damon_sysfs' features in the supported features
list to minimize the confusion.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>