_damon_dbgfs: Fix bug in test_init_regions_version()
In the test_init_regions_version() the original values of
debugfs_target_ids and debugfs_init_regions is saved so that it can be
re-written back at the end of the function.

The read_value_for_restore() returns a tuple with the value and the
error code. We need to keep these in two different variables as it is
done everywhere else in the code.

Change the call to read_value_for_restore() and get the error in a
seperate variable.

Signed-off-by: Puranjay Mohan <pjy@amazon.com>