_damon_result: Use PERF variable instead of 'perf' string
The perf tool is used in many places but there is no way to change the
path of 'perf' executable binary.

I have mistakenly modified only one of them to change the path of perf
binary, but it caused unexpected parsing error during damo record.

The problem happens because of using two different versions of 'perf'
binary one in 'perf record' and the other in 'perf script'.

To avoid such mistakes, this patch factors 'perf' string out as a global
variable PERF.

Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>