_damon/schemes_updated_running_kdamonds(): Return all kdamonds
Callers of schemes_updated_running_kdamonds() is not really wanting only
running kdamonds but all kdamonds that having updatd schemes stat and
tried_regions if those are running now.  Also, the function is returning
all kdamonds when the update was successful.  Rename the function for
better meaning, and fix return value when no kdamond is running.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>