Refactor damo_version.py
`damo_version.py` is currently a text file; this makes it difficult to
package in distributions which expect `*.py` files to be valid Python
that can be byte-compiled.

Convert it to a valid file, and adjust `damo` and `setup.py` to source
it appropriately. This also means we can stop patching the version in
with `sed` in `packaging/build.sh`.

Signed-off-by: Michel Alexandre Salim <salimma@fedoraproject.org>