_damon_args_schemes: Rename to _damo_deprecated
The single line per scheme input format code may need to exist a little
long, as 'translate_damos' subcommand should be supported for a while.
Rename _damon_args_schemes.py, which contains code for only a deprecated
feature to _damo_deprecated.py, to isolate the deprecated code while be
able to reused, and makes the meaning a bit more clear.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>