pre-commit: Introduce pre-commit hooks for automatic linting
This patch introduces pre-commit[1] to maintain the code quality with
automatic hooks execution.

This also applies the automatic changes by the pre-commit hooks.

More pre-commit hooks[2] can be added in the following patches and those
hooks will be added at .pre-commit-config.yaml.

The pre-commit tool can be downloaded and installed as follows.

  $ python3 -m pip install pre-commit

  $ cd /path/to/damo
  $ pre-commit install
  pre-commit installed at .git/hooks/pre-commit

Then pre-commit hooks will be automatically executed whenever a new
commit is created in the project.

[1] https://pre-commit.com
[2] https://pre-commit.com/hooks.html
Signed-off-by: Honggyu Kim <honggyu.kp@gmail.com>