_damon_result: Use '--force' for 'perf script'
There was a report[1] saying 'perf record' prints an error message
('failed to mmap file') and results in having the record output file
that having no proper ownership, in some setup.  It's unclear if it's
setup problem or perf's bug.  Nevertheless, it 'perf script' still works
with '--force' option.  For such cases, there is no reason for 'damo' to
be strict about any concern for 'perf'.  As long as it could parse, just
parse and use it.  For that, add '--force' option to the 'perf script'
command of 'damo'.

[1] https://github.com/awslabs/damo/issues/66

Signed-off-by: SeongJae Park <sj38.park@gmail.com>