_damo_fmt_str: Make text_to_{ratio,permil,percent}() to return more than 100% for 'max'
Ratios larger than 1.0 is possible, but text_to_{ratio,permil,percent}()
returns only {1.0,1000,100} for input 'max'.  This makes real issue when
a scheme is aimed to be set to capture all regions, and when a region
has larger than 100% access rate.  Such case is possible because
nr_accesses count logic is simple and rely on the time function, which
is not always strict.  Return ulong_max for 'max'.

Further update test_damo_scheme_dbgfs_conversion.py's expected outputs
for 'max' input together, to not break bisection.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>