USAGE: Remove 'damo schemes' section
'damo schemes' has no reason to exist since all its feature is available
with 'damo start'.  Only one unique feature of 'damo schemes' is waiting
for the termination of the kdamond.  Nevertheless, since it has
introduced in many public venues for demonstration, so would not easy to
deprecate.  Let's start by not talking about it anymore.  For that,
remove the section for 'damo schemes' from USAGE document.  The section
was anyway needed some update since some of the information such as type
of supported DAMOS actions is outdated.  The documentation of the list
would better to be done with some command line option, e.g.,
--damos_ls_actions?

Signed-off-by: SeongJae Park <sj38.park@gmail.com>