TODO: Remove memory efficient record file type item
We added the item because we found visualization report generation on my
laptop was consuming too much memory.  That was not due to ineffiency of
the json format, but because the damo on the local machine was not
updated to handle json format.  After updating the damo version on the
machine, it finishes the visualization report generation in reasonable
time, without OOM.  Remove the todo item.

Signed-off-by: SeongJae Park <sj@kernel.org>