_damon/read_feature_supports_file(): Support new format
The feature_supports file doesn't check if the saved content is for
debugfs interface or sysfs interface.  If a user does 'damo start' with
'--damon_interface sysfs' and then
'damo status --damon_interface debugfs', the output could be confusing.
Assume a new version of format, which saves the format version and the
feature supports separately for the interface.  Update reader first.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>