_damon_records: Retry only for DAMOS tried regions update failure
DAMOS treid regions update could fail due to conflicting operations,
e.g., when 'damo show' and 'damo status' are concurrently executed.
Currentl logic is checking the failure and retry up to five times with
exponential backoff.  But the scope of the retry is unnecessarily large.
Reduce the scope to really needed operation.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>