damo_show: Put new line between box and next line if max box height is 1
Boxes add one new line at the end to give bornder between boxes.  But
the new line is skipped if the max height is 1.  Hence, when max height
of box is 1 and there is something following the box on the format
string, outputs after box comes on same line.  Fix it.

Signed-off-by: SeongJae Park <sj38.park@gmail.com>