_damo_records: Add a new version of start_recording()
Currently wait_recording() is not only waiting, but also doing memory
footprint recording.  Let's make the user to create the handle with
requests part on their own, and start_recording() does all the real
works including waiting for completion.  As the first step of the work,
implement a new version of start_recording() function.

Signed-off-by: SeongJae Park <sj@kernel.org>