_damo_records: Poll target processes if exist
If 'damo record' is started with target command, the target process is
the child of 'damo', so becomes zombie process when it is finished.
Then because zombie process is still a valid process from DAMON's
perspective, kdamond is not terminated, and 'damo' hangs.  Do the pid
polling if any target process exists.

Signed-off-by: SeongJae Park <sj@kernel.org>