src/damo_report: Add a new format, heatmap
'heats' format is for raw heats data, to let users plot it in their way.
--heatmap option of the format does the plotting on behalf of users.
But it would be more convenient to do the plot by default.   To keep the
behavior same while making it more convenient, add another experimental
format 'heatmap'.  The format is just a copy of heats at the moment.

Signed-off-by: SeongJae Park <sj@kernel.org>