_damon: Handle a case that feature_supports file is written by an old version
From v2.0.4, damo caches kernel features supports checking results in
.damo.damon_feature_supports file.  That is for reducing unnecessary
complexities and IOs of the work, especially for asynchronous commands.
The cache file is saved by commands including start, record, and schemes
save the check, and loaded by commands including record (for online
recording case), show, and status.

In a corner case, user may saved the file with an old version of damo
(e.g., by damo start), and then loaded with a new version of damo (e.g.,
by damo status).  If the old version is oblivious of a feature while the
new version aware that, feature checks function could cause an
exception.

Fix this by checking if the features in the cached results is same to
the currently known list of features when reading the cache.

Reported-by: Honggyu Kim <honggyu.kim@sk.com>
Closes: https://github.com/awslabs/damo/issues/103
Signed-off-by: SeongJae Park <sj@kernel.org>