_damon_args: Implement --damos_filter_out
--damos_filter is confusing since the name doesn't explicitly saying if
it is doing filter-out or filter-in.  Due to the internal mechanism of
DAMON and allowance of multiple filters combinations, filter-out is more
preferred.  Add a similar but easy to use option, --damos_filter_out.
The name explains its behavior.  Further, the option allows 'matching'
argument be optional.  The default value is 'matching'.

Closes: https://github.com/damonitor/damo/issues/1
Signed-off-by: SeongJae Park <sj@kernel.org>